requirements.txt
aiohttp==3.0.7
async-timeout==2.0.0
attrs==17.4.0
beautifulsoup4==4.6.0
chardet==3.0.4
idna==2.6
idna-ssl==1.0.1
lxml==4.1.1
multidict==4.1.0
python-crontab==2.2.8
python-dateutil==2.6.1
pytz==2018.3
six==1.11.0
tzlocal==1.5.1
yarl==1.1.1
